requirements.txt
discord
youtube_dl
async_timeout
PyNaCl
pydantic
pyyaml